refactor battle code
