+++ src/web/src/common/helpers/utils.js
@@ -1,0 +1,8 @@
+export function delayCall(callback, duration) { 
+    setTimeout (
+      () => {
+        callback();
+      },
+      duration || 1000
+    );
+}
